+++ src/entities/EnemyActor.js
@@ -17,6 +17,8 @@ export class EnemyActor extends Phaser.Physics.Arcade.Sprite {
     this.weaponCooldowns = {};
     this.armorValue = options.armorValue ?? 0;
     this.baseTint = options.tint ?? null;
+    this.outlineTint = options.outlineTint ?? 0x6fb26d;
+    this.weaponVfxTint = options.weaponVfxTint ?? 0xf1cf58;
     this.canMove = true;
     this.baseY = y;
     this.onDeath = options.onDeath ?? null;
@@ -45,6 +47,10 @@ export class EnemyActor extends Phaser.Physics.Arcade.Sprite {
         0.2
       )
       .setDepth(3);
+    this.outlineRing = scene.add
+      .ellipse(this.x, this.y, this.isBoss ? 44 : 34, this.isBoss ? 44 : 34)
+      .setStrokeStyle(this.isBoss ? 4 : 3, this.outlineTint, 0.95)
+      .setDepth(3.5);
     this.healthBar = scene.add.graphics().setDepth(6);
     this.healthText = scene.add
       .text(this.x, this.y, "", {
@@ -91,6 +97,7 @@ export class EnemyActor extends Phaser.Physics.Arcade.Sprite {
 
   updatePresentation() {
     this.shadow.setPosition(this.x + 6, this.y + (this.isBoss ? 24 : 18));
+    this.outlineRing.setPosition(this.x, this.y);
     this.drawHealthBar();
   }
 
@@ -213,6 +220,10 @@ export class EnemyActor extends Phaser.Physics.Arcade.Sprite {
     if (this.shadow) {
       this.shadow.destroy();
       this.shadow = null;
+    }
+    if (this.outlineRing) {
+      this.outlineRing.destroy();
+      this.outlineRing = null;
     }
     if (this.healthBar) {
       this.healthBar.destroy();